Updated special_sort(array) function (V1.1.1
Added comments to clarify inner workings of the sorting function. A special sort function is required because the standard sort() function is tedious to work with when sorting duplicate key letters in subarrays.

Given [ ['A', 'Y'] and ['A', 'Z'] ], JS's sort() will sort the former subarray before the latter. This presents a problem if the latter subarray is the one that should start first.